requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
Flask-Bcrypt==1.0.1
Flask-CORS==4.0.0
python-dotenv==1.0.0
reportlab==4.0.4
pandas==2.1.1
pytest==7.4.2
pytest-flask==1.2.0
